requirements.txt
streamlit
duckdb
pandas
polars
pyyaml
python-dotenv